+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Reality Forger MK2 peripheral
+- Creative item duplicator
 
 ## [0.0.3] - 2021-07-29